requirements.txt
alembic==1.18.4
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
arviz==0.23.4
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cloudpickle==3.1.2
cons==0.4.7
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
databricks-sdk==0.102.0
distro==1.9.0
docker==7.1.0
etuples==0.3.10
fastapi==0.135.1
filelock==3.25.2
Flask==3.1.3
flask-cors==6.0.2
fonttools==4.62.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.1
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
greenlet==3.3.2
gunicorn==25.1.0
h11==0.16.0
h5netcdf==1.8.1
h5py==3.16.0
httpcore==1.0.9
httpx==0.28.1
huey==2.6.0
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
llvmlite==0.44.0
logical-unification==0.4.7
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
miniKanren==1.0.5
mlflow==3.10.1
mlflow-skinny==3.10.1
mlflow-tracing==3.10.1
multipledispatch==1.0.0
narwhals==2.18.0
numba==0.61.2
numpy==2.2.6
openai==2.26.0
opentelemetry-api==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
packaging==26.0
pandas==2.3.3
pillow==12.1.1
platformdirs==4.9.4
plotly==6.6.0
prettytable==3.17.0
protobuf==6.33.5
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pymc==5.28.1
pyparsing==3.3.2
pytensor==2.38.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
skops==0.13.0
smmap==5.0.3
sniffio==1.3.1
SQLAlchemy==2.0.48
sqlparse==0.5.5
starlette==0.52.1
streamlit==1.55.0
tenacity==9.1.4
threadpoolctl==3.6.0
toml==0.10.2
toolz==1.1.0
tornado==6.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.42.0
wcwidth==0.6.0
Werkzeug==3.1.6
xarray==2026.2.0
xarray-einstats==0.9.1
zipp==3.23.0
PyYAML==6.0.3
pytest==9.0.2
